fix: 修复16处Kotlin编译错误 - DanmakuManager/HttpUtils/DownloadManagerFactory/GestureController/AesCipherDataSource/BaseActivity
- DanmakuManager: DanmakuTimer.currMS 不可访问，改用 0L
- HttpUtils: inline函数访问private成员改internal，encodeToString显式传serializer
- DownloadManagerFactory: 构造参数类型不匹配，改用 DefaultDownloadIndex + DefaultDownloaderFactory
- GestureController: Float/Long 类型混用，统一转 Long
- AesCipherDataSource: TransferListener 非泛型，移除 <*> 类型参数
- BaseActivity: showLoading/hideLoading/navigateTo 改 internal 供 BaseFragment 调用

+++ app/src/main/java/com/zhuiju/app/core/player/GestureController.kt
@@ -137,7 +137,7 @@ class GestureController(
                 val percent = dx / totalWidth
                 val totalDuration = playerManager.progress.value.total
                 if (totalDuration > 0) {
-                    seekTargetMs = (seekTargetMs + percent * totalDuration).coerceIn(0, totalDuration)
+                    seekTargetMs = (seekTargetMs + (percent * totalDuration).toLong()).coerceIn(0L, totalDuration)
                     callback?.onSeekPreview(seekTargetMs, totalDuration)
                 }
             }
